# contracts/bridge/v1/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iceberg Platform Bridge",
    "description": "The stable interface for extension stacks. Authenticate with a Keycloak access token for the `iceberg-bridge` audience: a user token issued to the extension's clients, or the service-account token of the extension's confidential client. See contracts/bridge for the compatibility policy.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/bridge/v1/me": {
      "get": {
        "summary": "The signed-in user's teams, roles and databases",
        "operationId": "me_bridge_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContext"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "KeycloakBearer": []
          }
        ]
      }
    },
    "/bridge/v1/teams/{teamId}/automation-principals": {
      "get": {
        "summary": "This extension's automation principals of a team",
        "operationId": "team_automation_bridge_v1_teams__teamId__automation_principals_get",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^team-[a-f0-9]{32}$",
              "title": "Teamid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationList"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "KeycloakBearer": []
          }
        ]
      },
      "post": {
        "summary": "Enable this extension for a team environment (team administrators)",
        "operationId": "enable_bridge_v1_teams__teamId__automation_principals_post",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^team-[a-f0-9]{32}$",
              "title": "Teamid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationPrincipal"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationPrincipal"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "KeycloakBearer": []
          }
        ]
      }
    },
...
    "/bridge/v1/automation-principals": {
      "get": {
        "summary": "All automation principals of this extension",
        "operationId": "all_automation_bridge_v1_automation_principals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "KeycloakBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AutomationInput": {
        "properties": {
          "environment": {
            "type": "string",
            "enum": [
              "development",
              "acceptance",
              "production"
            ],
            "title": "Environment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "environment"
        ],
        "title": "AutomationInput"
      },
      "AutomationList": {
        "properties": {
          "automationPrincipals": {
            "items": {
              "$ref": "#/components/schemas/AutomationPrincipal"
            },
            "type": "array",
            "title": "Automationprincipals"
          }
        },
        "type": "object",
        "required": [
          "automationPrincipals"
        ],
        "title": "AutomationList"
      },
...
      "Error": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "title": "Error"
      },
...
      "Membership": {
        "properties": {
          "team": {
            "type": "string",
            "title": "Team"
          },
          "teamName": {
            "type": "string",
            "title": "Teamname"
          },
          "role": {
            "type": "string",
            "enum": [
              "reader",
              "writer",
              "admin",
              "bucket-admin"
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "team",
          "teamName",
          "role"
        ],
        "title": "Membership"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "User"
      },
      "UserContext": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/Membership"
            },
            "type": "array",
            "title": "Memberships"
          },
          "databases": {
            "items": {
              "$ref": "#/components/schemas/Database"
            },
            "type": "array",
            "title": "Databases"
          },
          "sharedDatabases": {
            "items": {
              "$ref": "#/components/schemas/SharedDatabase"
            },
            "type": "array",
            "title": "Shareddatabases"
          }
        },
        "type": "object",
        "required": [
          "user",
          "memberships",
          "databases",
          "sharedDatabases"
        ],
        "title": "UserContext"
      }
    },
    "securitySchemes": {
      "KeycloakBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak access token with audience iceberg-bridge."
      }
    }
  }
}